Refactor popup and close-popup templates to fix formatting discrepancies.
This commit refactors the popup template and close-popup template by improving the formatting to correct discrepancies. In the popup template, a div element and a footer element were added to the xpath while in the close-popup template, a style attribute was added to a div element to fix formatting differences.